// src/types/auth.ts
export interface IUser {
  _id: string;
  name: string;
  email: string;
  role: 'user' | 'admin';
  createdAt: string;
  updatedAt: string;
}

export interface AuthTokens {
  accessToken: string;
  refreshToken: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface AuthResponse {
  user: IUser;
  tokens: AuthTokens;
}

export interface JwtPayload {
  userId: string;
  email: string;
  role: 'user' | 'admin';
}